SPARK-48 moved the mechaclient to its own module
