+++ main.py
@@ -13,128 +13,17 @@ This module is built on top of the Pydle system.
 
 """
 import logging
-from typing import Callable
 
-from pydle import ClientPool, Client
+from pydle import ClientPool
 
 from Modules import rat_command
 from Modules.context import Context
-from Modules.event import Event
+from Modules.mechaclient import MechaClient
 from Modules.permissions import require_permission, RAT
 from Modules.rat_command import command
 from config import config
-from utils.ratlib import sanitize
 
 log = logging.getLogger(f"mecha.{__name__}")
-
-
-class MechaClient(Client):
-    """
-    MechaSqueak v3
-    """
-
-    __version__ = "3.0a"
-
-    def __init__(self, *args, **kwargs):
-        """
-        Custom mechasqueak constructor
-
-        Unused arguments are passed through to pydle's constructor
-
-        Args:
-            *args (list): arguments
-            **kwargs (list): keyword arguments
-
-        """
-        self._api_handler = None  # TODO: replace with handler init once it exists
-        self._database_manager = None  # TODO: replace with dbm once it exists
-        self._rat_cache = None  # TODO: replace with ratcache once it exists
-        super().__init__(*args, **kwargs)
-
-    @Event
-    async def on_connect(self):
-        """
-        Called upon connection to the IRC server
-        :return:
-        """
-        log.debug(f"Connecting to channels...")
-        # join a channel
-        for channel in config["irc"]["channels"]:
-            log.debug(f"Configured channel {channel}")
-            await self.join(channel)
-
-        log.debug("joined channels.")
-        for subscription in self._subscriptions["on_connect"]:
-            log.debug(f"invoking subscription for {subscription}")
-            subscription()
-        # call the super
-        super().on_connect()
-
-    #
-    # def on_join(self, channel, user):
-    #     super().on_join(channel, user)
-    @Event
-    async def on_message(self, channel, user, message: str):
-        """
-        Triggered when a message is received
-        :param channel: Channel the message arrived in
-        :param user: user that triggered the message
-        :param message: message body
-        :return:
-        """
-        log.debug(f"{channel}: <{user}> {message}")
-        log.debug("on_message invoking on_message_raw subscribers...")
-
-        for subscriber in self._subscriptions['on_message_raw']:
-            # call subscribers for the raw on_message event
-            logging.debug(f"invoking on_message_raw callback for {subscriber}...")
-            subscriber(channel=channel, sender=user, message=message)
-
-        if user == config['irc']['nickname']:
-            # don't do this and the bot can get into an infinite
-            # self-stimulated positive feedback loop.
-            log.debug(f"Ignored {message} (anti-loop)")
-            return None
-
-        if not message.startswith(rat_command.prefix):
-            # prevent bot from processing commands without the set prefix
-            log.debug(f"Ignored {message} (not a command)")
-
-            for subscriber in self._subscriptions['on_message']:
-                # call on_message subscriptions
-                await subscriber(channel=channel, sender=user, message=message)
-            return None
-
-        else:  # await command execution
-            # sanitize input string headed to command executor
-            sanitized_message = sanitize(message)
-            log.debug(f"Sanitized {sanitized_message}, Original: {message}")
-            for subscriber in self._subscriptions['on_command']:
-                await subscriber(message=sanitized_message, sender=user, channel=channel)
-            await rat_command.trigger(message=sanitized_message,
-                                      sender=user,
-                                      channel=channel)
-
-    @property
-    def rat_cache(self) -> object:
-        """
-        Mecha's rat cache
-        """
-        return self._rat_cache
-
-    @property
-    def database_mgr(self) -> object:
-        """
-        Mecha's database connection
-        """
-        return self._database_mgr
-
-    @property
-    def api_handler(self) -> object:
-        """
-        Mecha's API connection
-        """
-        return self._api_handler
 
 
 @require_permission(RAT)
